+++ packages/components-js/tests/e2e/specs/textarea-wrapper.e2e.ts
@@ -9,6 +9,7 @@ import {
   waitForStencilLifecycle,
   getElementInnerText,
   getElementStyle,
+  setAttribute,
 } from '../helpers';
 import { ElementHandle, Page } from 'puppeteer';
 import { FormState } from '@porsche-design-system/components/src/types';
@@ -27,9 +28,6 @@ describe('textarea-wrapper', () => {
   const getMessage = () => selectNode(page, 'p-textarea-wrapper >>> .message');
   const getLabel = () => selectNode(page, 'p-textarea-wrapper >>> .label__text');
   const getCounter = () => selectNode(page, 'p-textarea-wrapper >>> .counter');
-  const getLabelLink = () => selectNode(page, 'p-textarea-wrapper [slot="label"] a');
-  const getDescriptionLink = () => selectNode(page, 'p-textarea-wrapper [slot="description"] a');
-  const getMessageLink = () => selectNode(page, 'p-textarea-wrapper [slot="message"] a');
 
   type InitOptions = {
     useSlottedLabel?: boolean;
@@ -132,6 +130,17 @@ describe('textarea-wrapper', () => {
     await textarea.press('Backspace');
     await textarea.press('Backspace');
     expect(await getElementInnerText(counter)).toBe('0/160');
+  });
+
+  it('should render characterCountElement when maxlength is set', async () => {
+    await initTextarea();
+    const textarea = await getTextarea();
+
+    expect(await selectNode(page, 'p-textarea-wrapper >>> label .sr-only')).toBeNull();
+
+    await setAttribute(textarea, 'maxlength', '20');
+
+    expect(await selectNode(page, 'p-textarea-wrapper >>> label .sr-only')).toBeDefined();
   });
 
   describe('hover state', () => {