+++ packages/components/src/components/content/tabs/tabs/tabs-styles.spec.ts
@@ -1,7 +1,7 @@
 import { getComponentCss } from './tabs-styles';
 
 describe('getComponentCss()', () => {
-  it('should return correct css ', () => {
+  it('should return correct css', () => {
     expect(getComponentCss()).toMatchSnapshot();
   });
 });